% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipkg-package.R
\docType{package}
\name{epipkg-package}
\alias{epipkg}
\alias{epipkg-package}
\title{epipkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
